Dataset
The attached dataset has sensor data (accelerometer, magnetometer and gyroscope data) in conjunction with Google Activity labels (IN VEHICLE, WALKING, STILL). We used the dataset for building gait based models, where data rows corresponding to WALKING confidence ? 75% is extracted and analyzed. The code to analyze the data is not uploaded and Ping us if you need access to code to read and parse these files